+++ CHANGELOG.md
@@ -13,13 +13,14 @@
 
 ### Security
 
-## [3.3.0]
+## [4.0.0]
 ### Added
+
+- Kotlin Multiplatform support
 
 ### Changed
 
-- Update Android Gradle plugin to version 7+
-- Warning when license has no mapping for html reports
+- Extension property to set `enabled` and `destination`
 
 ### Deprecated
 
@@ -29,12 +30,22 @@
 
 ### Security
 
+## [3.3.0]
+
+### Changed
+
+- Update Android Gradle plugin to version 7+
+- Warning when license has no mapping for html reports
+
 ## [3.2.0]
+
 ### Added
+
 - set task outputs to report files
 - add DSL to configure reports
 
 ### Changed
+
 - use library to create CSV report
 
 ## [3.1.0]